Application restructure is correctly connected to redux & redux-form, still need to style
